+++ tas-installer/pkg/secrets/secret.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"securesign/sigstore-ocp/tas-installer/internal/kubernetes"
+	"securesign/sigstore-ocp/tas-installer/pkg/kubernetes"
 	"strings"
 	"time"
